requirements.txt
streamlit
streamlit[auth]

pandas
openpyxl
numpy

snowflake-connector-python
PyJWT[crypto]